Use long

+++ loss.py
@@ -66,7 +66,7 @@ class SigmoidFocalLoss(nn.Module):
         self.alpha = alpha
 
     def forward(self, out, target):
-        return nn.CrossEntropyLoss()(out, target)
+        return nn.CrossEntropyLoss()(out, target.long())
         '''
         n_class = out.shape[1]
         class_ids = torch.arange(
